inventory: Add number of items that has ...

+++ inventory-application/controllers/brand.js
@@ -1,7 +1,22 @@
 import Brand from "../models/Brand.js";
+import Perfume from "../models/Perfume.js";
 
+// TODO Learn how to use databases!
 export async function getAllBrands(req, res) {
-  const items = await Brand.find({});
+  const [perfumes, items] = await Promise.all([
+    Perfume.find({}).populate("brand"),
+    Brand.find({}),
+  ]);
+
+  const numbers = perfumes.reduce(
+    (acc, { brand: { name } }) => ({ ...acc, [name]: acc[name] + 1 || 1 }),
+    {}
+  );
+
+  for (const ele of items) {
+    ele.number = numbers?.[ele.name] ?? 0;
+  }
+
   res.render("brands", { items });
 }
 
@@ -38,7 +53,17 @@ export async function saveBrand(req, res) {
 export async function deleteBrand(req, res) {
   const { id } = req.params;
   try {
-    let brand = await Brand.findByIdAndDelete(id);
+    let perfumes = await Perfume.find({ brand: id });
+
+    if (perfumes !== []) {
+      let brand = await Brand.findById(id);
+      return res.render("brand", {
+        brand,
+        errors: ["There are perfumes of this brand"],
+      });
+    }
+
+    await Brand.findByIdAndDelete(id);
     res.redirect("/brands");
   } catch (error) {
     res.render("error", { error });
